requirements.txt
asgiref==3.8.1
Babel==2.14.0
Django==5.0.4
django-cors-headers==4.3.1
django-jazzmin==2.6.1
docxcompose==1.4.0
docxtpl==0.16.8
Jinja2==3.1.3
lxml==5.2.1
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
python-docx==1.1.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
